Rename scroling menu for user edit
